Moved HMM to base/estimator
