+++ aeon/base/estimator/hmm_learn/__init__.py
@@ -1,0 +1,8 @@
+"""Implements wrappers for estimators from hmmlearn."""
+
+__all__ = ["BaseHMMLearn", "GaussianHMM", "GMMHMM", "PoissonHMM"]
+
+from aeon.base.estimator.hmm_learn.base import BaseHMMLearn
+from aeon.base.estimator.hmm_learn.gaussian import GaussianHMM
+from aeon.base.estimator.hmm_learn.gmm import GMMHMM
+from aeon.base.estimator.hmm_learn.poisson import PoissonHMM
