+++ aeon/base/estimator/hmm_learn/gaussian.py
@@ -1,5 +1,5 @@
 """
-Hidden Markov Model with Poisson emissions.
+Hidden Markov Model with Gaussian emissions.
 
 Please see the original library
 (https://github.com/hmmlearn/hmmlearn/blob/main/lib/hmmlearn/hmm.py)
@@ -7,30 +7,48 @@ Please see the original library
 
 from typing import Dict
 
-from aeon.annotation.hmm_learn import BaseHMMLearn
+from aeon.base.estimator.hmm_learn import BaseHMMLearn
 
 __maintainer__ = []
-__all__ = ["PoissonHMM"]
+__all__ = ["GaussianHMM"]
 
 
-class PoissonHMM(BaseHMMLearn):
-    """
-    Hidden Markov Model with Poisson emissions.
+class GaussianHMM(BaseHMMLearn):
+    """Hidden Markov Model with Gaussian emissions.
 
     Parameters
     ----------
     n_components : int
-        Number of states.
+        Number of states
+    covariance_type : {"spherical", "diag", "full", "tied"}, optional
+        The type of covariance parameters to use:
+        * "spherical" --- each state uses a single variance value that
+            applies to all features.
+        * "diag" --- each state uses a diagonal covariance matrix
+            (default).
+        * "full" --- each state uses a full (i.e. unrestricted)
+            covariance matrix.
+        * "tied" --- all mixture components of each state use **the same**
+            full covariance matrix (note that this is not the same as for
+            `GaussianHMM`).
+    min_covar : float, optional
+        Floor on the diagonal of the covariance matrix to prevent
+        overfitting. Defaults to 1e-3.
+    means_prior, means_weight : array, shape (n_mix, ), optional
+        Mean and precision of the Normal prior distribtion for
+        :attr:`means_`.
+    covars_prior, covars_weight : array, shape (n_mix, ), optional
+        Parameters of the prior distribution for the covariance matrix
+        :attr:`covars_`.
+        If :attr:`covariance_type` is "spherical" or "diag" the prior is
+        the inverse gamma distribution, otherwise --- the inverse Wishart
+        distribution.
     startprob_prior : array, shape (n_components, ), optional
         Parameters of the Dirichlet prior distribution for
         :attr:`startprob_`.
     transmat_prior : array, shape (n_components, n_components), optional
         Parameters of the Dirichlet prior distribution for each row
         of the transition probabilities :attr:`transmat_`.
-    lambdas_prior, lambdas_weight : array, shape (n_components,), optional
-        The gamma prior on the lambda values using alpha-beta notation,
-        respectivley. If None, will be set based on the method of
-        moments.
     algorithm : {"viterbi", "map"}, optional
         Decoder algorithm.
     random_state: RandomState or an int seed, optional
@@ -47,8 +65,8 @@ class PoissonHMM(BaseHMMLearn):
     params, init_params : string, optional
         The parameters that get updated during (``params``) or initialized
         before (``init_params``) the training.  Can contain any
-        combination of 's' for startprob, 't' for transmat, and 'l' for
-        lambdas.  Defaults to all parameters.
+        combination of 's' for startprob, 't' for transmat, 'm' for
+        means and 'c' for covars.  Defaults to all parameters.
     implementation: string, optional
         Determines if the forward-backward algorithm is implemented with
         logarithms ("log"), or using scaling ("scaling").  The default is
@@ -56,54 +74,65 @@ class PoissonHMM(BaseHMMLearn):
 
     Attributes
     ----------
+    n_features : int
+        Dimensionality of the Gaussian emissions.
     monitor_ : ConvergenceMonitor
         Monitor object used to check the convergence of EM.
     startprob_ : array, shape (n_components, )
         Initial state occupation distribution.
     transmat_ : array, shape (n_components, n_components)
         Matrix of transition probabilities between states.
-    lambdas_ : array, shape (n_components, n_features)
-        The expectation value of the waiting time parameters for each
-        feature in a given state.
+    means_ : array, shape (n_components, n_features)
+        Mean parameters for each state.
+    covars_ : array
+        Covariance parameters for each state.
+        The shape depends on :attr:`covariance_type`:
+        * (n_components, )                        if "spherical",
+        * (n_components, n_features)              if "diag",
+        * (n_components, n_features, n_features)  if "full",
+        * (n_features, n_features)                if "tied".
 
     Examples
     --------
-    >>> from aeon.annotation.hmm_learn import PoissonHMM # doctest: +SKIP
-    >>> from from aeon.testing.utils.data_gen import piecewise_poisson # doctest: +SKIP
-    >>> data = piecewise_poisson( # doctest: +SKIP
-    ...    lambdas=[1, 2, 3], lengths=[2, 4, 8], random_state=7
+    >>> from aeon.annotation.hmm_learn import GaussianHMM # doctest: +SKIP
+    >>> from from aeon.testing.utils.data_gen import piecewise_normal # doctest: +SKIP
+    >>> data = piecewise_normal( # doctest: +SKIP
+    ...    means=[2, 4, 1], lengths=[10, 35, 40], random_state=7
     ...    ).reshape((-1, 1))
-    >>> model = PoissonHMM(n_components=3) # doctest: +SKIP
+    >>> model = GaussianHMM(algorithm='viterbi', n_components=2) # doctest: +SKIP
     >>> model = model.fit(data) # doctest: +SKIP
     >>> labeled_data = model.predict(data) # doctest: +SKIP
     """
 
-    _tags = {
-        "distribution_type": "Poisson",
-        # Tag to determine test for test_all_annotators
-    }
-
     def __init__(
         self,
         n_components: int = 1,
+        covariance_type: str = "diag",
+        min_covar: float = 1e-3,
         startprob_prior: float = 1.0,
         transmat_prior: float = 1.0,
-        lambdas_prior: float = 0.0,
-        lambdas_weight: float = 0.0,
+        means_prior: float = 0,
+        means_weight: float = 0,
+        covars_prior: float = 1e-2,
+        covars_weight: float = 1,
         algorithm: str = "viterbi",
-        random_state: int = None,
+        random_state: float = None,
         n_iter: int = 10,
         tol: float = 1e-2,
         verbose: bool = False,
-        params: str = "stl",
-        init_params: str = "stl",
+        params: str = "stmc",
+        init_params: str = "stmc",
         implementation: str = "log",
     ):
         self.n_components = n_components
+        self.covariance_type = covariance_type
+        self.min_covar = min_covar
         self.startprob_prior = startprob_prior
         self.transmat_prior = transmat_prior
-        self.lambdas_prior = lambdas_prior
-        self.lambdas_weight = lambdas_weight
+        self.means_prior = means_prior
+        self.means_weight = means_weight
+        self.covars_prior = covars_prior
+        self.covars_weight = covars_weight
         self.algorithm = algorithm
         self.random_state = random_state
         self.n_iter = n_iter
@@ -128,14 +157,18 @@ class PoissonHMM(BaseHMMLearn):
             Reference to self.
         """
         # import inside _fit to avoid hard dependency.
-        from hmmlearn.hmm import PoissonHMM as _PoissonHMM
+        from hmmlearn.hmm import GaussianHMM as _GaussianHMM
 
-        self._hmm_estimator = _PoissonHMM(
+        self._hmm_estimator = _GaussianHMM(
             self.n_components,
+            self.covariance_type,
+            self.min_covar,
             self.startprob_prior,
             self.transmat_prior,
-            self.lambdas_prior,
-            self.lambdas_weight,
+            self.means_prior,
+            self.means_weight,
+            self.covars_prior,
+            self.covars_weight,
             self.algorithm,
             self.random_state,
             self.n_iter,
@@ -163,4 +196,6 @@ class PoissonHMM(BaseHMMLearn):
         """
         return {
             "n_components": 3,
+            "covariance_type": "diag",
+            "min_covar": 1e-3,
         }
